+++ src/main/java/org/bedu/equipo15/inventary/service/serviceEquipment.java
@@ -1,13 +1,17 @@
 package org.bedu.equipo15.inventary.service;
 
 import org.bedu.equipo15.inventary.dto.*;
+import org.bedu.equipo15.inventary.exception.NotFoundException;
 import org.bedu.equipo15.inventary.mapper.mapperAddEquipment;
+import org.bedu.equipo15.inventary.mapper.mapperDepartament;
 import org.bedu.equipo15.inventary.mapper.mapperEquipment;
 import org.bedu.equipo15.inventary.model.Departament;
 import org.bedu.equipo15.inventary.model.Equipment;
+import org.bedu.equipo15.inventary.repository.repositoryDepartament;
 import org.bedu.equipo15.inventary.repository.repositoryEquipment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,29 +21,27 @@ public class serviceEquipment {
     @Autowired
     private mapperEquipment mapper;
     @Autowired
-    private mapperAddEquipment Addmapper;
+    private org.bedu.equipo15.inventary.mapper.mapperDepartament mapperDepartament;
     @Autowired
     private repositoryEquipment repository;
+    @Autowired
+    private org.bedu.equipo15.inventary.repository.repositoryDepartament repositoryDepartament;
 
     public List<dtoEquipment> findAll(){
 
         return repository.findAll().stream().map(mapper::toDTO).toList();
     }
 
-//    Pendiente Vincular campo Equipment. manufacturer = Manufacturer.id
-//    public dtoEquipment save(createpokemonDTO data){
-//        Equipment entity = repository.save(mapper.toModel(data));
-//        return mapper.toDTO(entity);
-//    }
+
     public dtoEquipment save(dtoEquipmentCreate data){
         Equipment entity = repository.save(mapper.toModel(data));
                 return mapper.toDTO(entity);
     }
-    public void update(long id, dtoUpdateEquipment data) {
+    public void update(long id, dtoUpdateEquipment data) throws NotFoundException {
         Optional<Equipment> result = repository.findById(id);
 
         if (result.isEmpty()) {
-            // throw new PacientNotFoundException(id);
+            throw new  NotFoundException(id);
         }
 
         Equipment model = result.get();
@@ -49,20 +51,43 @@ public class serviceEquipment {
         repository.save(model);
     }
 
-    public void deleteById(long id) {
+    public void deleteById(long id) throws NotFoundException {
+        if (!repository.existsById(id)) {
+            throw new NotFoundException(id);
+        }
+
         repository.deleteById(id);
     }
 
-    public void assignEquipment(long equipmentId, Departament departament) {
+    @Transactional
+    public void assignEquipment(long equipmentId, dtoDepartament departamentdto) throws NotFoundException {
         Optional<Equipment> equipmentOptional = repository.findById(equipmentId);
+
         if (equipmentOptional.isPresent()) {
-            Equipment model = equipmentOptional.get();
-            model.setDepartament(departament);
+            Optional<Departament> departamentOptional = repositoryDepartament.findById(departamentdto.getId());
 
-            // Acceder a la relación para evitar problemas de carga perezosa
-            departament.getEquipment().add(model);
+            if (departamentOptional.isPresent()) {
+                Equipment equipment = equipmentOptional.get();
 
-            repository.save(model);
+                // Map your DTO to the Departament entity using your mapper
+                Departament departament = mapperDepartament.toModelAdd(departamentdto);
+
+                // Save the departament
+                Departament savedDepartament = repositoryDepartament.save(departament);
+
+                // Set the relationship between departament and equipment
+                equipment.setDepartament(savedDepartament);
+
+                // Save the equipment entity
+                repository.save(equipment);
+            }else {
+                throw new NotFoundException(departamentdto.getId());
+            }
+
+        } else {
+
+            throw new NotFoundException(equipmentId);
+
         }
     }
 }
